refactor(PokerStore): rename files
